Pythonic renaming
